requirements.txt
kafka-python
flask
flask_restful